ref(IA): Organiza metodos de processamento das tags em classe
